api: update to 2.1.0 version
